+++ spec/rspec/rails/matchers/active_job_spec.rb
@@ -405,4 +405,262 @@ RSpec.describe "ActiveJob matchers", :skip => !RSpec::Rails::FeatureCheck.has_ac
         .to have_been_enqueued.at(a_value_within(5.seconds).of(future))
     end
   end
+
+  describe "have_performed_job" do
+    before do
+      ActiveJob::Base.queue_adapter.perform_enqueued_jobs = true
+      ActiveJob::Base.queue_adapter.perform_enqueued_at_jobs = true
+
+      # stub_const is used so `job_data["job_class"].constantize` works
+      stub_const('HeavyLiftingJob', heavy_lifting_job)
+      stub_const('HelloJob', hello_job)
+      stub_const('LoggingJob', logging_job)
+    end
+
+    it "raises ArgumentError when no Proc passed to expect" do
+      expect {
+        expect(heavy_lifting_job.perform_later).to have_performed_job
+      }.to raise_error(ArgumentError)
+    end
+
+    it "passes with default jobs count (exactly one)" do
+      expect {
+        heavy_lifting_job.perform_later
+      }.to have_performed_job
+    end
+
+    it "counts only jobs performed in block" do
+      heavy_lifting_job.perform_later
+      expect {
+        heavy_lifting_job.perform_later
+      }.to have_performed_job.exactly(1)
+    end
+
+    it "passes when negated" do
+      expect { }.not_to have_performed_job
+    end
+
+    it "fails when job is not performed" do
+      expect {
+        expect { }.to have_performed_job
+      }.to raise_error(/expected to perform exactly 1 jobs, but performed 0/)
+    end
+
+    it "fails when too many jobs performed" do
+      expect {
+        expect {
+          heavy_lifting_job.perform_later
+          heavy_lifting_job.perform_later
+        }.to have_performed_job.exactly(1)
+      }.to raise_error(/expected to perform exactly 1 jobs, but performed 2/)
+    end
+
+    it "reports correct number in fail error message" do
+      heavy_lifting_job.perform_later
+      expect {
+        expect { }.to have_performed_job.exactly(1)
+      }.to raise_error(/expected to perform exactly 1 jobs, but performed 0/)
+    end
+
+    it "fails when negated and job is performed" do
+      expect {
+        expect { heavy_lifting_job.perform_later }.not_to have_performed_job
+      }.to raise_error(/expected not to perform exactly 1 jobs, but performed 1/)
+    end
+
+    it "passes with job name" do
+      expect {
+        hello_job.perform_later
+        heavy_lifting_job.perform_later
+      }.to have_performed_job(hello_job).exactly(1).times
+    end
+
+    it "passes with multiple jobs" do
+      expect {
+        hello_job.perform_later
+        logging_job.perform_later
+        heavy_lifting_job.perform_later
+      }.to have_performed_job(hello_job).and have_performed_job(logging_job)
+    end
+
+    it "passes with :once count" do
+      expect {
+        hello_job.perform_later
+      }.to have_performed_job.exactly(:once)
+    end
+
+    it "passes with :twice count" do
+      expect {
+        hello_job.perform_later
+        hello_job.perform_later
+      }.to have_performed_job.exactly(:twice)
+    end
+
+    it "passes with :thrice count" do
+      expect {
+        hello_job.perform_later
+        hello_job.perform_later
+        hello_job.perform_later
+      }.to have_performed_job.exactly(:thrice)
+    end
+
+    it "passes with at_least count when performed jobs are over limit" do
+      expect {
+        hello_job.perform_later
+        hello_job.perform_later
+      }.to have_performed_job.at_least(:once)
+    end
+
+    it "passes with at_most count when performed jobs are under limit" do
+      expect {
+        hello_job.perform_later
+      }.to have_performed_job.at_most(:once)
+    end
+
+    it "generates failure message with at least hint" do
+      expect {
+        expect { }.to have_performed_job.at_least(:once)
+      }.to raise_error(/expected to perform at least 1 jobs, but performed 0/)
+    end
+
+    it "generates failure message with at most hint" do
+      expect {
+        expect {
+          hello_job.perform_later
+          hello_job.perform_later
+        }.to have_performed_job.at_most(:once)
+      }.to raise_error(/expected to perform at most 1 jobs, but performed 2/)
+    end
+
+    it "passes with provided queue name" do
+      expect {
+        hello_job.set(:queue => "low").perform_later
+      }.to have_performed_job.on_queue("low")
+    end
+
+    it "passes with provided at date" do
+      date = Date.tomorrow.noon
+      expect {
+        hello_job.set(:wait_until => date).perform_later
+      }.to have_performed_job.at(date)
+    end
+
+    it "passes with provided arguments" do
+      expect {
+        hello_job.perform_later(42, "David")
+      }.to have_performed_job.with(42, "David")
+    end
+
+    it "passes with provided arguments containing global id object" do
+      global_id_object = GlobalIdModel.new("42")
+
+      expect {
+        hello_job.perform_later(global_id_object)
+      }.to have_performed_job.with(global_id_object)
+    end
+
+    it "passes with provided argument matchers" do
+      expect {
+        hello_job.perform_later(42, "David")
+      }.to have_performed_job.with(42, "David")
+    end
+
+    it "generates failure message with all provided options" do
+      date = Date.tomorrow.noon
+      message = "expected to perform exactly 2 jobs, with [42], on queue low, at #{date}, but performed 0" + \
+                "\nQueued jobs:" + \
+                "\n  HelloJob job with [1], on queue default"
+
+      expect {
+        expect {
+          hello_job.perform_later(1)
+        }.to have_performed_job(hello_job).with(42).on_queue("low").at(date).exactly(2).times
+      }.to raise_error(message)
+    end
+
+    it "throws descriptive error when no test adapter set" do
+      queue_adapter = ActiveJob::Base.queue_adapter
+      ActiveJob::Base.queue_adapter = :inline
+
+      expect {
+        expect { heavy_lifting_job.perform_later }.to have_performed_job
+      }.to raise_error("To use ActiveJob matchers set `ActiveJob::Base.queue_adapter = :test`")
+
+      ActiveJob::Base.queue_adapter = queue_adapter
+    end
+
+    it "fails with with block with incorrect data" do
+      expect {
+        expect {
+          hello_job.perform_later("asdf")
+        }.to have_performed_job(hello_job).with { |arg|
+          expect(arg).to eq("zxcv")
+        }
+      }.to raise_error { |e|
+        expect(e.message).to match(/expected: "zxcv"/)
+        expect(e.message).to match(/got: "asdf"/)
+      }
+    end
+
+    it "passes multiple arguments to with block" do
+      expect {
+        hello_job.perform_later("asdf", "zxcv")
+      }.to have_performed_job(hello_job).with { |first_arg, second_arg|
+        expect(first_arg).to eq("asdf")
+        expect(second_arg).to eq("zxcv")
+      }
+    end
+
+    it "passess deserialized arguments to with block" do
+      global_id_object = GlobalIdModel.new("42")
+
+      expect {
+        hello_job.perform_later(global_id_object, :symbolized_key => "asdf")
+      }.to have_performed_job(hello_job).with { |first_arg, second_arg|
+        expect(first_arg).to eq(global_id_object)
+        expect(second_arg).to eq({:symbolized_key => "asdf"})
+      }
+    end
+
+    it "only calls with block if other conditions are met" do
+      noon = Date.tomorrow.noon
+      midnight = Date.tomorrow.midnight
+      expect {
+        hello_job.set(:wait_until => noon).perform_later("asdf")
+        hello_job.set(:wait_until => midnight).perform_later("zxcv")
+      }.to have_performed_job(hello_job).at(noon).with { |arg|
+        expect(arg).to eq("asdf")
+      }
+    end
+  end
+
+  describe "have_been_performed" do
+    before do
+      ActiveJob::Base.queue_adapter.performed_jobs.clear
+      ActiveJob::Base.queue_adapter.perform_enqueued_jobs = true
+      ActiveJob::Base.queue_adapter.perform_enqueued_at_jobs = true
+      stub_const('HeavyLiftingJob', heavy_lifting_job)
+    end
+
+    it "passes with default jobs count (exactly one)" do
+      heavy_lifting_job.perform_later
+      expect(heavy_lifting_job).to have_been_performed
+    end
+
+    it "counts all performed jobs" do
+      heavy_lifting_job.perform_later
+      heavy_lifting_job.perform_later
+      expect(heavy_lifting_job).to have_been_performed.exactly(2)
+    end
+
+    it "passes when negated" do
+      expect(heavy_lifting_job).not_to have_been_performed
+    end
+
+    it "fails when job is not performed" do
+      expect {
+        expect(heavy_lifting_job).to have_been_performed
+      }.to raise_error(/expected to perform exactly 1 jobs, but performed 0/)
+    end
+  end
 end